Formatter / Admin API / Migrate to Spring.
* Remove all jeeves services related to formatter admin
* Restore the functionnalities in the admin
* Hide 'important' formatter like xsl-view
* Preview / Add possibility to search for a record
* Add support to upload a formatter in a plugin - they are not all global.
