+++ web/src/main/webapp/WEB-INF/config.xml
@@ -156,7 +156,6 @@
   <include>config/config-service-admin.xml</include>
   <include>config/config-service-admin-harvesting.xml</include>
   <include>config/config-service-admin-schematron.xml</include>
-  <include>config/config-service-admin-formatter.xml</include>
   <include>config/config-service-csw.xml</include>
   <include>config/config-ui-search.xml</include>
   <include>config/config-ui-metadata.xml</include>
