--- web/src/main/webapp/WEB-INF/config/config-service-admin-formatter.xml
@@ -1,86 +1,0 @@
-<?xml version="1.0" encoding="ISO-8859-1"?>
-<!--
-  ~ Copyright (C) 2001-2016 Food and Agriculture Organization of the
-  ~ United Nations (FAO-UN), United Nations World Food Programme (WFP)
-  ~ and United Nations Environment Programme (UNEP)
-  ~
-  ~ This program is free software; you can redistribute it and/or modify
-  ~ it under the terms of the GNU General Public License as published by
-  ~ the Free Software Foundation; either version 2 of the License, or (at
-  ~ your option) any later version.
-  ~
-  ~ This program is distributed in the hope that it will be useful, but
-  ~ WITHOUT ANY WARRANTY; without even the implied warranty of
-  ~ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
-  ~ General Public License for more details.
-  ~
-  ~ You should have received a copy of the GNU General Public License
-  ~ along with this program; if not, write to the Free Software
-  ~ Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301, USA
-  ~
-  ~ Contact: Jeroen Ticheler - FAO - Viale delle Terme di Caracalla 2,
-  ~ Rome - Italy. email: geonetwork@osgeo.org
-  -->
-
-<geonet>
-  <services package="org.fao.geonet">
-    <service name="md.formatter.remove">
-      <documentation><![CDATA[
-        Remove a formatter.
-        ]]></documentation>
-      <class name=".api.records.formatters.Remove">
-        <param name="user_xsl_dir" value="user_xsl"/>
-      </class>
-      <error id="illegal-params" sheet="../xslt/error/error-xml.xsl"/>
-    </service>
-
-    <service name="md.formatter.download">
-      <documentation><![CDATA[
-        Download ZIP package containaing the formmatter.
-        ]]></documentation>
-      <class name=".api.records.formatters.Download">
-        <param name="user_xsl_dir" value="user_xsl"/>
-      </class>
-      <output file="true"/>
-      <error id="illegal-params" sheet="../xslt/error/error-xml.xsl"/>
-    </service>
-
-
-    <service name="md.formatter.files">
-      <documentation><![CDATA[
-        Return all files contained in a formatter package.
-        ]]></documentation>
-      <class name=".api.records.formatters.ListBundleFiles">
-        <param name="user_xsl_dir" value="user_xsl"/>
-      </class>
-      <error id="illegal-params" sheet="../xslt/error/error-xml.xsl"/>
-    </service>
-
-    <service name="md.formatter.edit">
-      <documentation><![CDATA[
-        Return a file contained in a formatter package.
-        ]]></documentation>
-      <class name=".api.records.formatters.EditFile">
-        <param name="user_xsl_dir" value="user_xsl"/>
-      </class>
-      <error id="illegal-params" sheet="../xslt/error/error-xml.xsl"/>
-    </service>
-
-    <service name="md.formatter.update">
-      <documentation><![CDATA[
-        Update a file in a formatter package.
-
-        Parameters:
-
-        * id: The formatter identifier
-        * fname: The file path in the formatter
-        * data: The file content
-        ]]></documentation>
-      <class name=".api.records.formatters.UpdateFile">
-        <param name="user_xsl_dir" value="user_xsl"/>
-      </class>
-      <error id="illegal-params" sheet="../xslt/error/error-xml.xsl"/>
-    </service>
-
-  </services>
-</geonet>
